requirements.txt
absl-py==1.3.0
aniso8601==9.0.1
apispec==6.0.2
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
flask-apispec==0.11.4
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SocketIO==5.3.2
flatbuffers==22.11.23
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
gunicorn==20.1.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.19.0
nltk==3.7
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
Pillow==9.3.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pymongo==4.3.3
python-dotenv==0.21.0
pytz==2022.6
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
spacy==3.3.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
tflearn==0.5.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
webargs==8.2.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
